Filtrado por fecha y carteles de confirmacion borrado y cargado

+++ client/src/context/gastosContext.js
@@ -1,4 +1,5 @@
 import { createContext, useEffect, useState } from "react";
+import Swal from "sweetalert2";
 
 const GastosContext = createContext();
 
@@ -23,14 +24,42 @@ const GastosProvider = ({children}) => {
     }, [actualizadoTabla])
 
     const deleteGasto = (id) => {
-        try{
-            fetch(`api/gastos/delete/${id}`, {
-                method: 'DELETE',
-                headers: {'Content-type': 'application/json; charset=UTF-8'}})
-            .then(() => setActualizadoTabla(true))
-        }catch(err){
-            console.log(err)
-        }
+        Swal.fire({
+            title: '¿Estás seguro de eliminar este gasto?',
+            showDenyButton: true,
+            confirmButtonText: 'Si',
+            denyButtonText: `No`,
+            background: '#198754',
+            color: '#fff',
+            confirmButtonColor: '#83d36d'
+        }).then((result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+                try{
+                    fetch(`api/gastos/delete/${id}`, {
+                        method: 'DELETE',
+                        headers: {'Content-type': 'application/json; charset=UTF-8'}})
+                    .then(() => setActualizadoTabla(true))
+                    Swal.fire({
+                        title: '¡Eliminado!',
+                        icon: 'success',
+                        timer: 1200,
+                        showConfirmButton: false,
+                    })
+                }catch(err){
+                    console.log(err)
+                }
+            } else if (result.isDenied) {
+                Swal.fire({
+                    title: 'Gasto no eliminado',
+                    icon: 'info',
+                    timer: 1200,
+                    timerProgressBar: true,
+                    showConfirmButton: false,
+                })
+            }
+        })
+        
     }
 
     const data = {
